fix: suppress noisy clang-tidy checks and address valid warnings
Disable clang-tidy checks that conflict with kernel conventions:
readability-redundant-declaration, bugprone-narrowing-conversions,
bugprone-assignment-in-if-condition, performance-no-int-to-ptr,
and readability-named-parameter. Enable clangd LSP integration.

Fix valid warnings across the codebase:
- Add VIRTIO_READ32/VIRTIO_WRITE32 macros for MMIO register access
- Fix macro argument parentheses and math precedence in headers
- Add explicit widening casts for multiplications used in 64-bit context
- Remove redundant casts and parentheses
- Split multi-declaration statements for readability
- Add missing default cases to switch statements
- Fix parameter name mismatches between declaration and definition
- Remove redundant file-scope extern declarations, restore function-scope ones

+++ arch/riscv/boot/mstart.c
@@ -30,7 +30,7 @@ __attribute__((noreturn)) void mstart(void)
 	sfence_vma();
 
 	int hartid = (int) r_mhartid();
-	w_mscratch((uint64) &mscratch0[hartid * 32]);
+	w_mscratch((uint64) &mscratch0[(int64) hartid * 32]);
 	w_mtvec((uint64) m_trap_handler);
 
 	uint64 x = r_mstatus();
